\name{FISHmQC-package}
\alias{FISHmQC-package}
\alias{FISHmQC}
\docType{package}
\title{
\packageTitle{FISHmQC}
}
\description{
\packageDescription{FISHmQC}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{FISHmQC}

Maintainer: \packageMaintainer{FISHmQC}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = FISHmQC)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{FISHmQC}}
%\section{Documentation Index}{\packageIndices{FISHmQC}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
